python/slack-bot/requirements.txt
Flask
slack_sdk
pyjwt
gql[all]
python-dotenv
